+++ requirements.txt
@@ -1,15 +1,21 @@
 annotated-types==0.7.0
 anyio==4.8.0
+blinker==1.9.0
 certifi==2025.1.31
 click==8.1.8
 colorama==0.4.6
 fastapi==0.115.11
+Flask==3.1.0
 h11==0.14.0
 httpcore==1.0.7
 httpx==0.28.1
 idna==3.10
+itsdangerous==2.2.0
+Jinja2==3.1.6
+MarkupSafe==3.0.2
 numpy==2.2.3
 ollama==0.4.7
+opencv-python==4.11.0.86
 pandas==2.2.3
 pydantic==2.10.6
 pydantic_core==2.27.2
@@ -22,4 +28,5 @@ starlette==0.46.1
 typing_extensions==4.12.2
 tzdata==2025.1
 uvicorn==0.34.0
+Werkzeug==3.1.3
 XlsxWriter==3.2.2
