--- templates/login.html
@@ -1,71 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" data-bs-theme="dark">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Login - ZoneWatch</title>
-
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="https://cdn.replit.com/agent/bootstrap-agent-dark-theme.min.css">
-    <!-- Font Awesome Icons -->
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css">
-
-    <!-- Firebase SDKs -->
-    <script src="https://www.gstatic.com/firebasejs/9.22.0/firebase-app-compat.js"></script>
-    <script src="https://www.gstatic.com/firebasejs/9.22.0/firebase-auth-compat.js"></script>
-
-    <!-- Auth Script -->
-    <script src="/static/js/auth.js"></script>
-</head>
-
-<body>
-    <div class="container mt-5">
-        <div class="row justify-content-center">
-            <div class="col-md-6">
-                <!-- Session expiry message (hidden by default) -->
-                <div id="sessionExpiredAlert" class="alert alert-warning mb-4 d-none">
-                    Your session has expired. Please sign in again.
-                </div>
-
-                <div class="card">
-                    <div class="card-body text-center">
-                        <h2 class="card-title mb-4">
-                            <i class="fas fa-video me-2"></i>ZoneWatch
-                        </h2>
-                        <p class="card-text mb-4">Please sign in to continue</p>
-                        <button id="signInBtn" class="btn btn-primary btn-lg">
-                            <i class="fab fa-google me-2"></i>Sign in with Google
-                        </button>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-
-    <script>
-        // Show session expired message if redirected from expired session
-        document.addEventListener('DOMContentLoaded', () => {
-            const params = new URLSearchParams(window.location.search);
-            if (params.get('expired') === 'true') {
-                document.getElementById('sessionExpiredAlert').classList.remove('d-none');
-            }
-        });
-
-        // Sign In with Google
-        document.getElementById('signInBtn').addEventListener('click', () => {
-            const provider = new firebase.auth.GoogleAuthProvider();
-            firebase.auth().signInWithPopup(provider)
-                .then((result) => {
-                    // Successful sign in, redirect to home page
-                    window.location.href = '/';
-                })
-                .catch((error) => {
-                    console.error('Error signing in:', error);
-                    alert('Error signing in: ' + error.message);
-                });
-        });
-    </script>
-</body>
-
-</html>
